a function that prints anything

+++ 0x10-variadic_functions/3-print_all.c
@@ -25,7 +25,7 @@ void print_all(const char * const format, ...)
 				check_stat = 0;
 				break;
 			case 'f':
-				printf("%d", va_arg(spc, int));
+				printf("%f", va_arg(spc, double));
 				check_stat = 0;
 				break;
 			case 'c':
